apps/listener-web-app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "PPP Listener Web App Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "firedantic",
    "firebase-admin",
    "uvicorn>=0.34.3",
    "google-adk>=1.4.2",
    "pydantic-async-validation[fastapi]>=0.3.0",
    "litellm>=1.72.4",
    "instructor>=1.8.3",
    "pydantic[email]>=2.11.5",
    "cryptography>=45.0.4",
    "google-cloud-aiplatform[adk,agent-engines]>=1.97.0",
    "pydantic-settings>=2.9.1",
    "aiohttp-sse-client>=0.2.1",
    "sseclient-py>=1.8.0",
    "httpx>=0.28.1",
    "httpx-sse==0.4.*",
    "google-cloud-logging>=3.12.1",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.34.1",
    "grpcio>=1.73.0",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "feedgen>=1.0.0",
    "fastmcp>=2.8.1",
    "fastapi-mcp>=0.3.4",
    "llama-index-core>=0.12.43"
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
lint = [
    "ruff>=0.12.0",
    "ty>=0.0.1a9",
]


[tool.ruff]
line-length = 200
include = ["*.py"]

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle, pyflakes, isort


[tool.ty]

[tool.ty.environment]
extra-paths = ["./ppp"]
python = "./.venv"
python-version = "3.12"


[tool.ty.src]
root = "."
exclude = ["tests"]
